+++ src/arena_tree.rs
@@ -33,6 +33,7 @@
 //
 //
 
+#![allow(dead_code)]
 
 use std::cell::Cell;
